vinh update lại cấu trúc project
