+++ src/main/java/com/mycompany/ungdungbanlaptop/repository/impl/ChucVuRepositoryImpl.java
@@ -1,0 +1,16 @@
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
+ */
+package com.mycompany.ungdungbanlaptop.repository.impl;
+
+import com.mycompany.ungdungbanlaptop.repository.ChucVuRepository;
+
+
+/**
+ *
+ * @author Diệm DZ
+ */
+public class ChucVuRepositoryImpl implements ChucVuRepository{
+     
+}
