+++ src/main/java/com/mycompany/ungdungbanlaptop/repository/impl/HoaDonChiTietRepositoryImpl.java
@@ -1,0 +1,84 @@
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
+ */
+package com.mycompany.ungdungbanlaptop.repository.impl;
+
+import com.mycompany.ungdungbanlaptop.entity.HoaDon;
+import com.mycompany.ungdungbanlaptop.entity.HoaDonChiTiet;
+import com.mycompany.ungdungbanlaptop.entity.SanPham;
+import com.mycompany.ungdungbanlaptop.repository.HoaDonChiTietRepository;
+import com.mycompany.ungdungbanlaptop.util.HibernateUtil;
+import java.math.BigDecimal;
+import java.util.List;
+import java.util.UUID;
+import org.hibernate.Session;
+import org.hibernate.Transaction;
+import org.hibernate.query.Query;
+
+/**
+ *
+ * @author Diệm DZ
+ */
+public class HoaDonChiTietRepositoryImpl implements HoaDonChiTietRepository {
+
+    @Override
+    public List<HoaDonChiTiet> getAll() {
+        try (Session session = HibernateUtil.getFACTORY().openSession()) {
+            Query query = session.createQuery("FROM SanPham");
+            List<HoaDonChiTiet> list = query.getResultList();
+
+            return list;
+        } catch (Exception e) {
+            e.printStackTrace(System.out);
+        }
+        return null;
+    }
+
+    @Override
+    public HoaDonChiTiet add(HoaDonChiTiet hoaDonChiTiet) {
+        Transaction transaction = null;
+        try (Session session = HibernateUtil.getFACTORY().openSession()) {
+            transaction = session.beginTransaction();
+            session.save(hoaDonChiTiet);
+            transaction.commit();
+
+        } catch (Exception e) {
+            e.printStackTrace(System.out);
+        }
+        return hoaDonChiTiet;
+    }
+
+    @Override
+    public HoaDonChiTiet update(HoaDonChiTiet hoaDonChiTiet) {
+        Transaction transaction = null;
+        try (Session session = HibernateUtil.getFACTORY().openSession()) {
+            transaction = session.beginTransaction();
+            session.update(hoaDonChiTiet);
+            transaction.commit();
+
+        } catch (Exception e) {
+            e.printStackTrace(System.out);
+        }
+        return hoaDonChiTiet;
+    }
+
+    @Override
+    public HoaDonChiTiet delete(HoaDonChiTiet hoaDonChiTiet) {
+        Transaction transaction = null;
+        try (Session session = HibernateUtil.getFACTORY().openSession()) {
+            transaction = session.beginTransaction();
+            session.delete(hoaDonChiTiet);
+            transaction.commit();
+
+        } catch (Exception e) {
+            e.printStackTrace(System.out);
+        }
+        return hoaDonChiTiet;
+    }
+
+    public static void main(String[] args) {
+//        HoaDonChiTiet hdct = new HoaDonChiTiet(1, BigDecimal.valueOf(Double.valueOf(2000)), new HoaDon(UUID.fromString("c0a83801-847e-19b9-8184-7ea9be380000")));
+//        System.out.println(new HoaDonChiTietRepositoryImpl().add(hdct));
+    }
+}
