+++ src/main/java/com/mycompany/ungdungbanlaptop/service/impl/BaoHanhServiceImpl.java
@@ -6,6 +6,7 @@ package com.mycompany.ungdungbanlaptop.service.impl;
 
 import com.mycompany.ungdungbanlaptop.entity.BaoHanh;
 import com.mycompany.ungdungbanlaptop.repository.BaoHanhRepository;
+import com.mycompany.ungdungbanlaptop.repository.impl.BaoHanhRepositoryImpl;
 import com.mycompany.ungdungbanlaptop.service.BaoHanhService;
 import java.util.List;
 
@@ -14,7 +15,7 @@ import java.util.List;
  * @author Du
  */
 public class BaoHanhServiceImpl implements BaoHanhService{
-    private BaoHanhRepository ql = new BaoHanhRepository();
+    private BaoHanhRepository ql = new BaoHanhRepositoryImpl();
 
     @Override
     public List<BaoHanh> getAll() {
