+++ src/main/java/com/mycompany/ungdungbanlaptop/service/impl/CPUServiceImpl.java
@@ -7,7 +7,8 @@ package com.mycompany.ungdungbanlaptop.service.impl;
 import com.mycompany.ungdungbanlaptop.entity.CPU;
 import com.mycompany.ungdungbanlaptop.entity.NhanVien;
 import com.mycompany.ungdungbanlaptop.repository.CPURespository;
-import com.mycompany.ungdungbanlaptop.repository.NhanVienRepository;
+import com.mycompany.ungdungbanlaptop.repository.impl.CPURespositoryImpl;
+import com.mycompany.ungdungbanlaptop.repository.impl.NhanVienRepositoryImpl;
 import com.mycompany.ungdungbanlaptop.service.CPUService;
 import java.util.List;
 
@@ -17,7 +18,7 @@ import java.util.List;
  */
 public class CPUServiceImpl implements CPUService {
 
-    private CPURespository cpuRespository = new CPURespository();
+    private CPURespository cpuRespository = new CPURespositoryImpl();
 
     @Override
     public String add(CPU cpu) {
